requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
iniconfig==2.1.0
    # via pytest
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
pydantic==2.11.7
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pyswisseph==2.10.3.2
    # via -r requirements.in
pytest==8.4.1
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
